Move `api/internal/rawhelpers.go` to `internal`
