+++ internal/rawhelpers.go
